seperate out fgamma and fk
